+++ InstantMock/Arguments/Argument/ArgumentVerify/ArgumentVerify.swift
@@ -1,0 +1,11 @@
+//
+//  ArgumentVerify.swift
+//  InstantMock
+//
+//  Created by Patrick on 12/05/2017.
+//  Copyright © 2017 pirishd. All rights reserved.
+//
+
+
+/** Protocol for an argument that must verify a precise condition */
+public protocol ArgumentVerify: Argument {}