+++ InstantMock/Arguments/Helpers/ArgumentFactory.swift
@@ -17,12 +17,15 @@ public protocol ArgumentFactory {
     /// Create a new any argument
     func argumentAny(_ typeDescription: String) -> ArgumentAny
 
+    // Create a new argument closure
+    func argumentClosure(_ typeDescription: String) -> ArgumentClosure
+
     // Create a new argument that must verify provided condition
     func argument(condition: @escaping (Value) -> Bool) -> ArgumentVerify
+    func argument(condition: @escaping (Value?) -> Bool) -> ArgumentVerify
 
     // Create a new argument capture
     func argumentCapture(_ typeDescription: String) -> ArgumentCapture
-
 }
 
 
@@ -37,8 +40,16 @@ class ArgumentFactoryImpl<T>: ArgumentFactory {
         return ArgumentAnyImpl(typeDescription)
     }
 
+    func argumentClosure(_ typeDescription: String) -> ArgumentClosure {
+        return ArgumentClosureImpl(typeDescription)
+    }
+
     func argument(condition: @escaping (T) -> Bool) -> ArgumentVerify {
-        return ArgumentVerifyImpl<T>(condition)
+        return ArgumentVerifyMandatoryImpl<T>(condition)
+    }
+
+    func argument(condition: @escaping (T?) -> Bool) -> ArgumentVerify {
+        return ArgumentVerifyOptionalImpl<T>(condition)
     }
 
     func argumentCapture(_ typeDescription: String) -> ArgumentCapture {