+++ InstantMock/Interceptors/Expectation.swift
@@ -43,7 +43,7 @@ public class Expectation: CallInterceptor {
 
     /** Method is being called */
     @discardableResult
-    override func handleCall(_ args: [Any?]) -> Any? {
+    override func handleCall(_ args: [Any?]) throws -> Any? {
         self.numberOfCalls = self.numberOfCalls + 1
         return nil // don't care about return values
     }
@@ -57,8 +57,8 @@ extension Expectation {
 
     /** register call */
     @discardableResult
-    public func call<T>(_ value: T, numberOfTimes: Int? = nil) -> Stub {
-        self.expectedNumberOfCalls = numberOfTimes
+    public func call<T>(_ value: T, count: Int? = nil) -> Stub {
+        self.expectedNumberOfCalls = count
         return self.stub
     }