+++ InstantMockTests/Integration Tests/ClosureMockTests.swift
@@ -15,13 +15,18 @@ class SomeClosureObject {}
 
 
 protocol ClosureProtocol {
-    func someFunc(arg: String, closure: ((_ arg1: String, _ arg2: SomeClosureObject) -> Int))
+    func someFunc(arg: String, closure: (_ arg1: String, _ arg2: SomeClosureObject) -> Int)
+    func someFuncOpt(arg: String?, closure: ((_ arg1: String, _ arg2: SomeClosureObject) -> Int)?)
 }
 
 
 class ClosureMock: Mock, ClosureProtocol {
 
-    func someFunc(arg: String, closure: ((String, SomeClosureObject) -> Int)) {
+    func someFunc(arg: String, closure: (String, SomeClosureObject) -> Int) {
+        super.call(arg, closure)
+    }
+
+    func someFuncOpt(arg: String?, closure: ((String, SomeClosureObject) -> Int)?) {
         super.call(arg, closure)
     }
 
@@ -45,8 +50,8 @@ class ClosureMockTests: XCTestCase {
     func testExpect() {
 
         self.mock.expect().call(self.mock.someFunc(
-            arg: Arg<String>.any,
-            closure: Arg<Closure>.any.cast() as (String, SomeClosureObject) -> Int
+            arg: Arg.any(),
+            closure: Arg.closure()
         ))
 
         self.mock.someFunc(arg: "Hello", closure: { (str, obj) -> Int in
@@ -57,5 +62,37 @@ class ClosureMockTests: XCTestCase {
         XCTAssertTrue(self.assertionMock.succeeded)
 
     }
+
+
+    func testExpect_optionalNil() {
+
+        self.mock.expect().call(self.mock.someFuncOpt(
+            arg: Arg.any(),
+            closure: Arg.closure()
+        ))
+
+        self.mock.someFuncOpt(arg: "Hello", closure: nil)
+
+        self.mock.verify()
+        XCTAssertTrue(self.assertionMock.succeeded)
+
+    }
+
+
+    func testExpect_optionalNotNil() {
+
+        self.mock.expect().call(self.mock.someFuncOpt(
+            arg: Arg.any(),
+            closure: Arg.closure()
+        ))
+
+        self.mock.someFuncOpt(arg: "Hello", closure: { (str, obj) -> Int in
+            return 42
+        })
+
+        self.mock.verify()
+        XCTAssertTrue(self.assertionMock.succeeded)
+    }
+
 }