+++ InstantMockTests/Unit Tests/Interceptors/ExpectationTests.swift
@@ -24,7 +24,7 @@ class ExpectationTests: XCTestCase {
 
 
     func testHandleCall() {
-        let ret = self.expectation.handleCall([])
+        let ret = try! self.expectation.handleCall([])
         XCTAssertNil(ret)
     }
 
@@ -36,26 +36,26 @@ class ExpectationTests: XCTestCase {
 
 
     func testVerified_called() {
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         var verified = self.expectation.verified
         XCTAssertTrue(verified)
 
         // call a second time
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         verified = self.expectation.verified
         XCTAssertTrue(verified)
     }
 
 
     func testVerified_withExpectedNumberOfCalls() {
-        self.expectation.call(Int.any, numberOfTimes: 2)
+        self.expectation.call(Int.any, count: 2)
 
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         var verified = self.expectation.verified
         XCTAssertFalse(verified)
 
         // call a second time
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         verified = self.expectation.verified
         XCTAssertTrue(verified)
     }
@@ -77,12 +77,12 @@ class ExpectationTests: XCTestCase {
         let config = CallConfiguration(for: "Func", with: self.argsConfig)
         self.expectation.configuration = config
 
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         var reason = self.expectation.reason
         XCTAssertNil(reason)
 
         // call a second time
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         reason = self.expectation.reason
         XCTAssertNil(reason)
     }
@@ -92,14 +92,14 @@ class ExpectationTests: XCTestCase {
         let config = CallConfiguration(for: "Func", with: self.argsConfig)
         self.expectation.configuration = config
 
-        self.expectation.call(Int.any, numberOfTimes: 2)
+        self.expectation.call(Int.any, count: 2)
 
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         var reason = self.expectation.reason
         XCTAssertEqual(reason, "Func not called the expected number of times (1 out of 2) with expected args (none)")
 
         // call a second time
-        self.expectation.handleCall([])
+        try! self.expectation.handleCall([])
         reason = self.expectation.reason
         XCTAssertNil(reason)
     }