+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+
+## 0.2.0
+
+* Simplify syntax for `Arg.any()`
+
+
+## 0.1.0
+
+* Add `andThrow` to stubs
+
+
+## 0.0.5
+
+* Fix public visibility of initializers for `ArgumentCaptor` and `ArgumentClosureCaptor`
+
+
+## 0.0.4
+
+* Fix optional values in `Arg.verify`
+* Make syntax for handling closures simpler
+* Make syntax for number of expected calls simpler
+
+
+## 0.0.3
+
+* Fix API for `verify()` to take into account #file and #line
+* Fix optional values in `Arg.eq`
+
+
+## 0.0.2
+
+* Attempt to generate doc
+
+
 ## 0.0.1
 
 * Say Hello to InstantMock